rename: 변경된 컴포넌트 이름을 Tabs2 -> Tabs으로 변경하고 import 수정
