+++ FE/src/components/foundation/Tabs/TabPanel.tsx
@@ -1,16 +1,16 @@
 import { HTMLElementTypes } from '@/types/utils';
-import useTabs from '@foundation/Tabs2/useTabs';
+import useTabs from '@foundation/Tabs/useTabs';
 
 type TabPanelProps = {
   children: React.ReactNode;
   value: string;
 } & HTMLElementTypes<HTMLDivElement>;
 
-const TabPanel2: React.FC<TabPanelProps> = ({ children, value, ...args }) => {
+const TabPanel: React.FC<TabPanelProps> = ({ children, value, ...args }) => {
   const { currentValue } = useTabs();
 
   if (currentValue !== value) return null;
   return <div {...args}>{children}</div>;
 };
 
-export default TabPanel2;
+export default TabPanel;
