+++ FE/src/components/foundation/Tabs/useTabs.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { TabContext } from '@foundation/Tabs2/index';
+import { TabContext } from '@foundation/Tabs/index';
 
 const useTabs = () => {
   return useContext(TabContext);
